% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calibration.R
\name{BMIQcalibration}
\alias{BMIQcalibration}
\title{Calibrate methylation data to a gold standard.}
\usage{
BMIQcalibration(datM)
}
\arguments{
\item{datM}{A matrix of beta values that have been cleaned
with \code{\link{clean_beta}}, with samples in columns and
CpGs in rows.}
}
\value{
A matrix of beta-values calibrated to the gold standard GSE50498.
}
\description{
\code{BMIQcalibration} uses an adapted version of the BMIQ algorithm to
calibrate DNA methylation data from Illumina HumanMethylation arrays to
the gold standard dataset used in the muscle clock (GSE50498).
}
\details{
\code{BMIQcalibration} was created by Steve Horvath,
largely based on the \code{\link[wateRmelon]{BMIQ}} function from
Teschendorff 2013 to adjust for the type-2 bias in Illumina HM450
and HMEPIC data. BMIQ stands for beta mixture quantile normalization.
Horvath fixed minor errors in the v_1.2 version of the BMIQ algorithm
and changed the optimization algorithm to make the code more robust.
He used method = "Nelder-Mead" in \code{\link[stats]{optim}} since
the other optimization method sometimes gets stuck. Toward this end,
the function \code{\link[RPMM]{blc}} was replaced by \code{blc2}.
\code{datM} needs to be a matrix of beta-values that have been
cleaned and reduced to the same CpGs as the gold standard,
using the \code{\link{clean_beta}} function.
Each sample in \code{datM} is iteratively calibrated to the
gold standard values, so the time it takes to run
\code{BMIQcalibration} is directly proportional to the number
of samples in \code{datM}. This step is essential to estimate
epigenetic age with accuracy.
}
\examples{
data("GSE121961", envir = environment())
GSE121961_clean <- clean_beta(beta = GSE121961)
GSE121961_calibrated <- BMIQcalibration(datM = GSE121961_clean)

}
\seealso{
\code{\link{clean_beta}} to get the DNA methylation matrix ready
for calibration,
\code{\link[wateRmelon]{BMIQ}} for the original BMIQ algorithm and
\url{https://genomebiology.biomedcentral.com/articles/10.1186/gb-2013-14-10-r115}
for the original paper describing Horvath's adapted BMIQ algorithm.
}
