% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticnet_model_MEAT2.0.R
\docType{data}
\name{elasticnet_model_MEAT2.0}
\alias{elasticnet_model_MEAT2.0}
\title{Elastic net model used in the updated muscle clock (MEAT 2.0).}
\format{
An elastic net model
}
\usage{
elasticnet_model_MEAT2.0
}
\description{
An object with S3 class "glmnet","elnet" generated by training 1,053
skeletal muscle DNA methylation profiles on a tranformed version of age.
This elastic net model can take in any skeletal muscle DNA methylation
profile that has been cleaned and calibrated to the GSE50498 gold standard
dataset, to estimate epigenetic age in the sample.
}
\seealso{
\code{\link[glmnet]{glmnet}}
}
\keyword{datasets}
