% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiage_estimation.R
\name{epiage_estimation}
\alias{epiage_estimation}
\title{Estimates age in skeletal muscle from calibrated DNA methylation profiles.}
\usage{
epiage_estimation(beta, pheno = NULL, ID_col_name = NULL, age_col_name = NULL)
}
\arguments{
\item{beta}{A matrix with beta values that have been calibrated to
the gold standard GSE50498 using \code{\link{BMIQcalibration}},
with samples in columns and CpGs in rows.}

\item{pheno}{An optional data frame, matrix or tibble containing information
on the samples such as age, sex, disease status, etc.}

\item{ID_col_name}{The name of the column in \code{pheno} that contains
the sample IDs corresponding to the column names of \code{beta}.}

\item{age_col_name}{The name of the column in \code{pheno} that contains
age (in years).}
}
\value{
A list containing the following items:
\enumerate{
  \item \code{DNAmage} A numeric vector of predicted age (in years)
  for each sample in \code{beta} based on DNA methylation profiles.
  \item \code{AAdiff} An optional numeric vector corresponding to the
  difference between predicted and actual age (in years) for each sample
  in \code{beta}. This is only returned if \code{age_col_name} is not null.
  \item \code{AAresid} An optional numeric vector corresponding to
  the residuals of linear model (using \code{\link[stats]{lm}}) of
  predicted against actual age  for each sample in \code{beta}.
  This is only returned if \code{age_col_name} is not null.
}
}
\description{
\code{epiage_estimation} takes as input a calibrated methylation matrix of
beta-values and uses the muscle clock to estimate age.
}
\details{
\code{epiage_estimation} estimates age for each sample in the input
\code{beta} DNA methylation matrix based on DNA methylation profiles.
\code{beta} needs to be a matrix of beta-values that have been calibrated
to the gold standard GSE50498 using \code{\link{BMIQcalibration}}.
}
\examples{
data("GSE121961", envir = environment())
GSE121961_clean <- clean_beta(beta = GSE121961)
GSE121961_calibrated <- BMIQcalibration(datM = GSE121961_clean)

# Example without specifying phenotypes
# (output = predicted age only)
GSE121961_epiage <- epiage_estimation(beta = GSE121961_calibrated)

# Example with phenotypes
# (output = predicted age,
# age acceleration difference and age acceleration residual)
data("GSE121961_pheno", envir = environment())
GSE121961_epiage <- epiage_estimation(
  beta = GSE121961_calibrated,
  pheno = GSE121961_pheno,
  ID_col_name = "ID",
  age_col_name = "Age"
)
}
\seealso{
\code{\link[wateRmelon]{BMIQ}} for the original BMIQ algorithm,
\url{https://genomebiology.biomedcentral.com/articles/10.1186/gb-2013-14-10-r115}
for the adapted version of the BMIQ algorithm, and
\url{https://genomebiology.biomedcentral.com/articles/10.1186/gb-2013-14-10-r115}
for the elastic net model of the muscle clock.
}
