% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE121961.R
\docType{data}
\name{GSE121961}
\alias{GSE121961}
\title{GSE121961 methylation data}
\format{A data frame with 866,091 CpGs (rows) and 2 individuals (columns)}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE121nnn/GSE121961/matrix/}
}
\usage{
GSE121961
}
\description{
GSE121961 dataset containing 2 DNA methylation profiles generated with
the HMEPIC technology, and used here as a test dataset.
}
\keyword{datasets}
