% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE121961_pheno.R
\docType{data}
\name{GSE121961_pheno}
\alias{GSE121961_pheno}
\title{GSE121961 phenotypes}
\format{A data frame with 2 samples (rows) and 4 phenotypes (columns).}
\source{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/humu.23745}
}
\usage{
GSE121961_pheno
}
\description{
GSE121961_pheno contains information on sex,
age (missing for the controls),
and group (Control, or SELENON/RYR mutant)
for the 2 samples in the GSE121961 DNA methylation dataset.
}
\keyword{datasets}
