% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_beta}
\alias{clean_beta}
\title{Clean beta-matrix.}
\usage{
clean_beta(beta)
}
\arguments{
\item{beta}{A matrix or data frame of DNA methylation beta-values,
with samples in columns and CpGs in rows.}
}
\value{
A clean version of the input \code{beta} reduced to 19,401 CpGs,
with missing values imputed, and without 0 or 1 values.
}
\description{
\code{clean_beta} reduces \code{beta} to the right CpGs, imputes missing
values if any, and replaces 0 and 1 with min and max values.
}
\details{
\code{clean_beta} will
1) reduce \code{beta} to the 19,401 CpGs used
to calibrate DNA methylation profiles to the gold standard,
2) check whether \code{beta} contains missing values,
and impute them with \code{\link[impute]{impute.knn}},
3) check whether \code{beta} contains 0 and 1 values, and if any,
change them to the minimum non-0 and maximum non-1 values in \code{beta}.
}
\examples{
data("GSE121961", envir = environment())
GSE121961_clean <- clean_beta(beta = GSE121961)
}
\seealso{
\code{\link[impute]{impute.knn}} for imputation of missing values.
}
