% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEAT-package.R
\docType{package}
\name{MEAT-package}
\alias{MEAT}
\alias{MEAT-package}
\title{MEAT: Muscle Epigenetic Age Test}
\description{
This package estimates epigenetic age in skeletal muscle,
    using DNA methylation data generated with the
    Illumina Infinium technology (HM27, HM450 and HMEPIC).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sarah-voisin/MEAT}
  \item Report bugs at \url{https://github.com/sarah-voisin/MEAT/issues}
}

}
\author{
\strong{Maintainer}: Sarah Voisin \email{sarah.voisin.aeris@gmail.com} (\href{https://orcid.org/0000-0002-4074-7083}{ORCID})

Other contributors:
\itemize{
  \item Steve Horvath \email{SHorvath@mednet.ucla.edu} (\href{https://orcid.org/0000-0002-4110-3589}{ORCID}) [contributor]
}

}
\keyword{internal}
